refactor: decompose FloorPlanDesigner into modular sub-components
Extracted the 1,491-line FloorPlanDesigner monolith into:
- src/constants/floorPlanPresets.js (PRESETS + MOCK_ATTENDEES)
- src/utils/floorPlanGeometry.js (checkCollision, getSeatPositions)
- src/utils/floorPlanExport.js (export/import helpers)
- src/components/events/FloorPlan/ElementPalette.jsx (left toolbox)
- src/components/events/FloorPlan/PropertiesPanel.jsx (right panel)

FloorPlanDesigner.js now acts as a lean orchestrator importing from these modules.

Closes #5511
