+++ src/constants/floorPlanPresets.js
@@ -1,0 +1,32 @@
+export const MOCK_ATTENDEES = [
+  "Amit Sharma", "Priya Singh", "Rohit Verma", "Neha Kapoor",
+  "Vikram Rathore", "Siddharth Malhotra", "Kriti Sanon", "Varun Dhawan",
+  "Aditi Rao", "Ranbir Kapoor", "Deepika Padukone", "Ranveer Singh",
+  "Alia Bhatt", "Ayushmann Khurrana", "Rajkummar Rao", "Shraddha Kapoor"
+];
+
+export const PRESETS = {
+  empty: [],
+  banquet: [
+    { id: "stage-1", type: "stage", label: "Main Stage", x: 350, y: 50, width: 300, height: 120, rotation: 0, seatsCount: 0, assignedAttendees: {} },
+    { id: "table-1", type: "round-table", label: "VIP Table A", x: 200, y: 300, width: 140, height: 140, rotation: 0, seatsCount: 8, assignedAttendees: { 0: "Amit Sharma", 1: "Priya Singh" } },
+    { id: "table-2", type: "round-table", label: "VIP Table B", x: 660, y: 300, width: 140, height: 140, rotation: 0, seatsCount: 8, assignedAttendees: { 2: "Rohit Verma", 3: "Neha Kapoor" } },
+    { id: "table-3", type: "round-table", label: "Table 1", x: 120, y: 520, width: 120, height: 120, rotation: 0, seatsCount: 6, assignedAttendees: {} },
+    { id: "table-4", type: "round-table", label: "Table 2", x: 440, y: 520, width: 120, height: 120, rotation: 0, seatsCount: 6, assignedAttendees: {} },
+    { id: "table-5", type: "round-table", label: "Table 3", x: 760, y: 520, width: 120, height: 120, rotation: 0, seatsCount: 6, assignedAttendees: {} },
+    { id: "booth-1", type: "booth", label: "Sound & Lights", x: 450, y: 750, width: 100, height: 80, rotation: 0, seatsCount: 0, assignedAttendees: {} },
+    { id: "barrier-1", type: "barrier", label: "Security Line", x: 250, y: 220, width: 500, height: 10, rotation: 0, seatsCount: 0, assignedAttendees: {} }
+  ],
+  conference: [
+    { id: "stage-1", type: "stage", label: "Keynote Stage", x: 300, y: 60, width: 400, height: 130, rotation: 0, seatsCount: 0, assignedAttendees: {} },
+    { id: "podium-1", type: "booth", label: "Presenter Stand", x: 480, y: 210, width: 40, height: 40, rotation: 0, seatsCount: 0, assignedAttendees: {} },
+    { id: "rows-1", type: "rect-table", label: "Front Row A", x: 150, y: 340, width: 300, height: 60, rotation: 0, seatsCount: 10, assignedAttendees: {} },
+    { id: "rows-2", type: "rect-table", label: "Front Row B", x: 550, y: 340, width: 300, height: 60, rotation: 0, seatsCount: 10, assignedAttendees: {} },
+    { id: "rows-3", type: "rect-table", label: "Middle Row A", x: 150, y: 460, width: 300, height: 60, rotation: 0, seatsCount: 10, assignedAttendees: {} },
+    { id: "rows-4", type: "rect-table", label: "Middle Row B", x: 550, y: 460, width: 300, height: 60, rotation: 0, seatsCount: 10, assignedAttendees: {} },
+    { id: "rows-5", type: "rect-table", label: "Back Row A", x: 150, y: 580, width: 300, height: 60, rotation: 0, seatsCount: 10, assignedAttendees: {} },
+    { id: "rows-6", type: "rect-table", label: "Back Row B", x: 550, y: 580, width: 300, height: 60, rotation: 0, seatsCount: 10, assignedAttendees: {} },
+    { id: "exit-1", type: "exit", label: "Main Exit", x: 50, y: 800, width: 80, height: 20, rotation: 0, seatsCount: 0, assignedAttendees: {} },
+    { id: "exit-2", type: "exit", label: "Emergency Exit", x: 870, y: 800, width: 80, height: 20, rotation: 0, seatsCount: 0, assignedAttendees: {} }
+  ]
+};
